farm_back/requirements.txt
amqp==2.4.2
Babel==2.7.0
celery==4.3.0
certifi==2019.3.9
cffi==1.15.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.2.1
defusedxml==0.6.0
Django==2.2
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-cors-headers==2.4.0
django-filter==2.4.0
django-rest-swagger==2.2.0
django-rosetta==0.9.0
django-simple-history==2.7.3
django-storages==1.7.1
django-timezone-field==3.0
djangorestframework==3.9.0
djangorestframework-gis==0.14
docutils==0.14
drf-yasg==1.15.0
et-xmlfile==1.0.1
flower==0.9.3
html5lib==1.0.1
idna==2.8
imagesize==1.1.0
inflection==0.3.1
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.5.0
MarkupSafe==1.1.1
mock==3.0.5
numpy==1.23.0
openapi-codec==1.3.2
openpyxl==2.6.2
Pillow==9.1.1
pluggy==0.12.0
polib==1.1.0
pycparser==2.19
Pygments==2.4.2
pyparsing==2.4.2
Pyphen==0.9.5
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
ruamel.yaml==0.17.21
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tinycss2==1.0.2
tornado==5.1.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.2
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
zipp==0.5.2
